refactor(fuse-vhd): move to @vates namespace
