+++ @vates/fuse-vhd/.npmignore
