+++ @vates/fuse-vhd/index.js
